Custom Modhook and Exception emails
Added a custom modhook that implements a few hooks to help with user
management. E.g. when a player joins or when a player's data gets saved

Added a way for mod developers to obtain exceptions from the game by
having json data sent to their email with the data they pass through
and what exception occured to help with debugging a problem. Though
this can lead to spam if a error occurs multiple times so may need to
make a filter system of some kind. But for now will send all exceptions
to the mod devs. Will require feedback on how errors should be handled