Fixed issue where error logs was causing errors
Moved error logs for mods to the root folder of staxel as they were
causing hash issues for the bundler. This is something I will need
to look out for in the future. Maybe offer an entity to exist in the
world to pass settings for mods between the server and client.

I believe this will save on errors and offer an easy pathway for setting
updates without affecting the bundler