Moved configs and errors
To stop future hash conflicts the core mod will now store configs and
errors in the root folder rather than the content folder which would
cause issues after a content build and the files had changed in some way

Added in a settings manager which will send settings for mods to the
client and allow them to be read. This occurs when a player joins
or when a setting has been updated. This simplifies the need for an
extra channel to send configs via the content bundler and allows the
settings to be updated dynamically on the client when updated on the
server